Added product quantity to ProductDetailScreen.
